pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "byteflow"
version = "0.1.0"
description = "A lightweight PyTorch image classification training framework"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
authors = [
    { name = "Med Vall Cheikh" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0"
]

[tool.setuptools.packages.find]
include = ["byteflow", "byteflow.*"]
